+++ test/scour/filter_test.js
@@ -2,62 +2,25 @@
 
 const scour = require('../../scour')
 
-const data = {
-  users: {
-    1: { name: 'john' },
-    2: { name: 'jake' },
-    3: { name: 'ara' }
-  }
-}
+describe('.filter() (simple)', function () {
+  it('works for objects/functions', function () {
+    const data = { a: 10, b: 11, c: 12, d: 13 }
+    const result = scour(data).filter((item) => item.value % 2 === 0)
 
-describe('.filter()', function () {
-  beforeEach(function () {
-    this.results = scour(data).go('users').filter({ name: { $regex: /^j/ } })
+    expect(result.value).toEqual({ a: 10, c: 12 })
   })
 
-  it('works', function () {
-    expect(this.results.go(1).value).toEqual({ name: 'john' })
-    expect(this.results.go(2).value).toEqual({ name: 'jake' })
-    expect(this.results.go(3)).toEqual(undefined)
+  it('works for objects/objects', function () {
+    const data = { a: { num: 10 }, b: { num: 11 }, c: { num: 12 } }
+    const result = scour(data).filter({ num: 11 })
+
+    expect(result.value).toEqual({ b: { num: 11 } })
   })
 
-  describe('on objects', function () {
-    beforeEach(function () {
-      this.results = scour(data).go('users').filter({ name: { $regex: /^a/ } })
-    })
+  it('works for objects/queries', function () {
+    const data = { a: { num: 10 }, b: { num: 11 }, c: { num: 12 } }
+    const result = scour(data).filter({ num: { $eq: 11 } })
 
-    it('has results indexed by id', function () {
-      expect(this.results.go(3).value).toEqual({ name: 'ara' })
-    })
-
-    it('has proper keypaths', function () {
-      expect(this.results.go(3).keypath).toEqual([ 'users', '3' ])
-    })
-  })
-
-  describe('on empties', function () {
-    beforeEach(function () {
-      this.results = scour(data).go('users').filter({ abc: 'def' })
-    })
-
-    it('works', function () {
-      expect(this.results.len()).toEqual(0)
-    })
-  })
-})
-
-describe('.filter() via function', function () {
-  it('works for objects', function () {
-    var data = { a: 10, b: 11, c: 12 }
-    var result = scour(data).filter((val, key) => val.value % 2 === 0)
-
-    expect(result).toEqual({ a: 10, c: 12 })
-  })
-
-  it('works for arrays', function () {
-    var data = [ 10, 11, 12 ]
-    var result = scour(data).filter((val, key) => val.value % 2 === 0)
-
-    expect(result).toEqual([ 10, 12 ])
+    expect(result.value).toEqual({ b: { num: 11 } })
   })
 })